Add cmake dependency to trigger regeneration of the message header after the initial build
Without this dependency, if a plugin header has been added after the initial build of a message
package, it would not be picked up even if a cmake invocation is enforced with the `--force-cmake`
argument or manually because cmake considers the generated header as up-to-date.

The patch also adds some cmake output to indicate that a user-defined message or plugin header has
been found in the package source.